+++ e-httpRequest/src/main/java/eu/unifiedviews/plugins/extractor/httprequest/HttpRequestConfig_V1.java
@@ -1,12 +1,18 @@
 package eu.unifiedviews.plugins.extractor.httprequest;
 
+import eu.unifiedviews.helpers.dpu.ontology.EntityDescription;
+import eu.unifiedviews.plugins.extractor.httprequest.rdfConfig.FormParamBody;
+
 import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
 import java.util.Map;
 
+@EntityDescription.Entity(type = HttpRequestVocabulary.STR_CONFIG_CLASS)
 public class HttpRequestConfig_V1 {
 
     public enum DataType {
-        RAW_DATA, FORM_DATA, FILE;
+        RAW_DATA, FORM_DATA, FILE, FORM_DATA_RDF;
     }
 
     public enum RequestType {
@@ -21,8 +27,10 @@ public class HttpRequestConfig_V1 {
 
     private RequestContentType contentType = RequestContentType.TEXT;
 
+    @EntityDescription.Property(uri = HttpRequestVocabulary.STR_URL)
     private String requestURL = "";
 
+    @EntityDescription.Property(uri = HttpRequestVocabulary.STR_RAW_REQUEST_BODY)
     private String rawRequestBody = "";
 
     private String userName;
@@ -33,9 +41,18 @@ public class HttpRequestConfig_V1 {
 
     private Map<String, String> formDataRequestBody = new HashMap<>();
 
-    private static final String FILE_NAME_DEFAULT = "http_response";
+    private String fileName = "http_response";
 
-    private String fileName = FILE_NAME_DEFAULT;
+    @EntityDescription.Property(uri = HttpRequestVocabulary.STR_RAW_FORM_PARAMS_BODIES)
+    private List<FormParamBody> formParamBodies = new LinkedList<>();
+
+    public List<FormParamBody> getFormParamBodies() {
+        return formParamBodies;
+    }
+
+    public void setFormParamBodies(List<FormParamBody> formParamBodies) {
+        this.formParamBodies = formParamBodies;
+    }
 
     public DataType getPostRequestDataType() {
         return this.postRequestDataType;