+++ e-httpRequest/src/main/java/eu/unifiedviews/plugins/extractor/httprequest/HttpRequestHelper.java
@@ -1,11 +1,11 @@
 package eu.unifiedviews.plugins.extractor.httprequest;
 
-import java.io.IOException;
-
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.impl.client.CloseableHttpClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
 
 public class HttpRequestHelper {