+++ e-sparqlEndpoint/src/main/java/eu/unifiedviews/plugins/extractor/sparqlendpoint/RemoteRdfDataUnit.java
@@ -1,18 +1,5 @@
 package eu.unifiedviews.plugins.extractor.sparqlendpoint;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.ListIterator;
-import java.util.Set;
-
-import org.openrdf.model.URI;
-import org.openrdf.repository.RepositoryConnection;
-import org.openrdf.repository.RepositoryException;
-import org.openrdf.repository.sparql.SPARQLRepository;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import eu.unifiedviews.dataunit.DataUnitException;
 import eu.unifiedviews.dataunit.rdf.RDFDataUnit;
 import eu.unifiedviews.dpu.DPUException;
@@ -21,6 +8,14 @@ import eu.unifiedviews.helpers.dpu.context.Context;
 import eu.unifiedviews.helpers.dpu.exec.ExecContext;
 import eu.unifiedviews.helpers.dpu.extension.Extension;
 import eu.unifiedviews.helpers.dpu.extension.ExtensionException;
+import  org.eclipse.rdf4j.model.IRI;
+import  org.eclipse.rdf4j.repository.RepositoryConnection;
+import  org.eclipse.rdf4j.repository.RepositoryException;
+import  org.eclipse.rdf4j.repository.sparql.SPARQLRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.*;
 
 /**
  * Represents remote RDF repository in form of {@link RDFDataUnit}.
@@ -48,7 +43,7 @@ public class RemoteRdfDataUnit implements RDFDataUnit, AutoCloseable, Extension,
 
     private boolean closed = false;
 
-    RemoteRdfDataUnit(ExecContext<?> execContext, String endpointUrl, URI... graphs)
+    RemoteRdfDataUnit(ExecContext<?> execContext, String endpointUrl, IRI... graphs)
             throws ExternalError {
         this.execContext = execContext;
         this.repository = new SPARQLRepository(endpointUrl);
@@ -59,7 +54,7 @@ public class RemoteRdfDataUnit implements RDFDataUnit, AutoCloseable, Extension,
         }
         // Initialize entries.
         entries = new ArrayList<>(graphs.length);
-        for (URI uri : graphs) {
+        for (IRI uri : graphs) {
             entries.add(new RdfDataUnitUtils.InMemoryEntry(uri, "remote: " + endpointUrl));
         }
     }
@@ -97,7 +92,7 @@ public class RemoteRdfDataUnit implements RDFDataUnit, AutoCloseable, Extension,
     }
 
     @Override
-    public Set<URI> getMetadataGraphnames() throws DataUnitException {
+    public Set<IRI> getMetadataGraphnames() throws DataUnitException {
         return Collections.EMPTY_SET;
     }
 
@@ -126,6 +121,7 @@ public class RemoteRdfDataUnit implements RDFDataUnit, AutoCloseable, Extension,
         if (execPoint == ExecutionPoint.POST_EXECUTE) {
             // If RdfDataUnit is not closed then call close.
             if (!closed) {
+                LOG.debug("RdfDatunit not closed, calling close()");
                 try {
                     close();
                 } catch (Exception ex) {