+++ e-sparqlEndpoint/src/main/java/eu/unifiedviews/plugins/extractor/sparqlendpoint/SparqlEndpointVaadinDialog.java
@@ -1,8 +1,8 @@
 package eu.unifiedviews.plugins.extractor.sparqlendpoint;
 
-import org.openrdf.query.MalformedQueryException;
-import org.openrdf.query.QueryLanguage;
-import org.openrdf.query.parser.QueryParserUtil;
+import  org.eclipse.rdf4j.query.MalformedQueryException;
+import  org.eclipse.rdf4j.query.QueryLanguage;
+import  org.eclipse.rdf4j.query.parser.QueryParserUtil;
 
 import com.vaadin.data.Validator;
 import com.vaadin.ui.AbstractSelect.NewItemHandler;