+++ t-rdfValidator/src/main/java/eu/unifiedviews/plugins/transformer/rdfvalidator/RdfValidatorVaadinDialog.java
@@ -1,11 +1,11 @@
 package eu.unifiedviews.plugins.transformer.rdfvalidator;
 
-import org.openrdf.query.MalformedQueryException;
-import org.openrdf.query.QueryLanguage;
-import org.openrdf.query.parser.ParsedBooleanQuery;
-import org.openrdf.query.parser.ParsedQuery;
-import org.openrdf.query.parser.ParsedTupleQuery;
-import org.openrdf.query.parser.QueryParserUtil;
+import org.eclipse.rdf4j.query.MalformedQueryException;
+import org.eclipse.rdf4j.query.QueryLanguage;
+import org.eclipse.rdf4j.query.parser.ParsedBooleanQuery;
+import org.eclipse.rdf4j.query.parser.ParsedQuery;
+import org.eclipse.rdf4j.query.parser.ParsedTupleQuery;
+import org.eclipse.rdf4j.query.parser.QueryParserUtil;
 
 import com.vaadin.data.Validator;
 import com.vaadin.data.util.ObjectProperty;