+++ t-tabular/src/main/java/eu/unifiedviews/plugins/transformer/tabular/column/ValueGeneratorString.java
@@ -2,9 +2,9 @@ package eu.unifiedviews.plugins.transformer.tabular.column;
 
 import java.util.List;
 
-import org.openrdf.model.URI;
-import org.openrdf.model.Value;
-import org.openrdf.model.ValueFactory;
+import org.eclipse.rdf4j.model.IRI;
+import org.eclipse.rdf4j.model.Value;
+import org.eclipse.rdf4j.model.ValueFactory;
 
 import cz.cuni.mff.xrg.uv.transformer.tabular.column.ValueGeneratorReplace;
 
@@ -20,7 +20,7 @@ public class ValueGeneratorString extends ValueGeneratorReplace {
      */
     private final String language;
 
-    public ValueGeneratorString(URI uri, String template, String language) {
+    public ValueGeneratorString(IRI uri, String template, String language) {
         super(uri, template);
         this.language = language;
     }