+++ e-distributionMetadata/CHANGELOG.md
@@ -1,6 +1,10 @@
 E-DistributionMetadata
 ----------
 
+v3.0.0
+---
+* Updated to use UnifiedViews helpers 3.X with rdf4j support
+
 v2.0.5
 ---
 * improvements in slovak translations